Update README.md
To create a Python program that shows different charts of the S&P 500 over the last 20 years, you'll need historical data for the S&P 500. You can obtain this data from sources like Yahoo Finance or Alpha Vantage. In this example, we'll use the Yahoo Finance API to retrieve the data and visualize it using the Pandas and Matplotlib libraries.

Make sure you have the necessary libraries installed. If not, install them using pip: